+++ Bot/Bot/Models/StrategyResult.cs
@@ -5,12 +5,12 @@ namespace Bot.Models;
 
 class StrategyResult
 {
-    public List<(DateTime, double)> PortfolioValue { get; set; }
+    public List<(DateTime, decimal)> PortfolioValue { get; set; }
 
     /// <summary>
     /// Adds the value for a particular datetime.
     /// </summary>
-    public void AddPortfolioValue(DateTime dateTime, double value)
+    public void AddPortfolioValue(DateTime dateTime, decimal value)
     {
         PortfolioValue.Add((dateTime, value));
     }
