+++ Bot/FormRunner/ScatterPlotForm.cs
@@ -2,7 +2,7 @@ namespace FormRunner
 {
     public partial class ScatterPlotForm : Form
     {
-        public ScatterPlotForm(DateTime[] xData, double[] yData)
+        public ScatterPlotForm(DateTime[] xData, decimal[] yData)
         {
             InitializeComponent();
             formsPlot1.Plot.Add.Scatter(xData.Select(d => d.ToOADate()).ToArray(), yData);
